refactor: default trait methods on Emitter trait to reduce boilerplate
